Only one pop-up window instead of two

+++ main.py
@@ -13,7 +13,6 @@ USERNAME = os.getenv("SPOTIFY_USERNAME")  # username for Spotify account
 AMOUNT_OF_SONGS = 20
 
 
-
 def scrape_data(input_date):
     # scraping info about top artist from billboard page for given date
     scraper = BillboardScraper(date=input_date)
@@ -43,10 +42,11 @@ if __name__ == "__main__":
 
     app = PlaylistCreatorGUI()
     app.mainloop()
-    username, input_date = app.playlist_created()
-    artist_name, artist_name_endpoint, artist_img, artist_songs, path_to_img = scrape_data(input_date=input_date)
-    create_playlist(name=artist_name, songs=artist_songs,
-                    image=artist_img)
-    # wait 3 seconds (for playlist being created) and delete image from images folder
-    time.sleep(3)
-    path_to_img.unlink()
+    if app.name and app.date:
+        # username, input_date = app.playlist_created()
+        artist_name, artist_name_endpoint, artist_img, artist_songs, path_to_img = scrape_data(input_date=app.date)
+        create_playlist(name=artist_name, songs=artist_songs,
+                        image=artist_img)
+        # wait 3 seconds (for playlist being created) and delete image from images folder
+        time.sleep(3)
+        path_to_img.unlink()
